Moves Float Market Types to Separate File
